Rollback planned storage changes if partitioning method changes
The partitioning tools might be able plan storage changes the
other tools might not understand.
So just rollback the changes for now when the user switches
between the methods on the storage spoke while showing a warning
in the info bar. Like this users can cancell the switch to
not loose their planned partitioning changes.