Hide storage config spokes marked by SAM as visited
If the custom partitioning spoke or the Blivet GUI spoke
is marked as visited by the Screen Access Manager then
hide it's radio button.

This is mainly expected to be used to hide the custom storage
spoke or Blivet GUI spoke (mostly likely not both) in custom
Fedora spins that want to provide only one non-automatic storage
configuration option.